requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
google-generativeai==0.5.4
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
